Update Check GoogleStorage.Service.deleteURL

+++ src/google-storage/google-storage.service.ts
@@ -212,9 +212,15 @@ export class GoogleStorageService {
     input: InputDeleteFileOnStorage,
   ): Promise<{ message: string }> {
     try {
+      console.log(input);
+      if (!input.fileName.includes('storage.googleapis.com')) {
+        return { message: 'URL file is not from google storage' };
+      }
       const bucket = this.getBucket();
       const parts = input.fileName.split('/');
       const fileName = parts.slice(4).join('/');
+
+      console.log(fileName);
       bucket
         .file(fileName)
         .delete()
